requirements.txt
asgiref==3.10.0
Django==5.2.7
django-jazzmin==3.0.1
django-phonenumber-field==8.4.0
pillow==12.0.0
python-dotenv==1.2.1
sqlparse==0.5.3
phonenumbers-9.0.20
